requirements.txt
Django==3.2.14
python-telegram-bot==13.0
